LayoutTests/http/wpt/wk-web-animations/timing-model/keyframe-effect-interface-timing-duration.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Check that KeyframeEffect objects expose a duration through its timing property</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

internals.pauseTimeline(document.timeline);

test(t => {
    const keyframeEffect = new KeyframeEffect(document.body);
    assert_inherits(keyframeEffect, "timing");
    assert_inherits(keyframeEffect.timing, "duration");
    assert_equals(keyframeEffect.timing.duration, 0);
}, "Check the default state of the timing and duration properties.");

test(t => {
    const keyframeEffect = new KeyframeEffect(document.body);
    keyframeEffect.timing.duration = 2;
    assert_equals(keyframeEffect.timing.duration, 2);
}, "Check that we can update the timing duration.");

</script>
</body>
